// generated/clients/javascript/services/uigf-uncategorized-ugc/src/models/models.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

import { NodeReadableStream } from "@azure/core-rest-pipeline";

export type LevelApiGetCollectionTinyInfoResponse = {
  /**
   * BROWSER ONLY
   *
   * The response body as a browser Blob.
   * Always `undefined` in node.js.
   */
  blobBody?: Promise<Blob>;
  /**
   * NODEJS ONLY
   *
   * The response body as a node.js Readable stream.
   * Always `undefined` in the browser.
   */
  readableStreamBody?: NodeReadableStream;
};

export type LevelApiGetIndexRedpointResponse = {
  /**
   * BROWSER ONLY
   *
   * The response body as a browser Blob.
   * Always `undefined` in node.js.
   */
  blobBody?: Promise<Blob>;
  /**
   * NODEJS ONLY
   *
   * The response body as a node.js Readable stream.
   * Always `undefined` in the browser.
   */
  readableStreamBody?: NodeReadableStream;
};